+++ src/Vitorm/Entity/PropertyType/ETypeMode.cs
@@ -1,0 +1,12 @@
+﻿namespace Vitorm.Entity.PropertyType
+{
+    public enum ETypeMode
+    {
+        value,
+        @object,
+        array,
+        dictionary,
+
+    }
+
+}